requirements.txt
azure-functions
azure-iot-hub
requests
pandas
numpy